chore: move tests/compliance/ to test/compliance/ for consistency
Stage 3 introduced tests/ (plural) alongside the existing test/ (singular)
without a reason — the existing convention is everything under test/.
Rename to consolidate. No content changes.

- git mv tests/compliance test/compliance
- Updated import paths in moved test files (../../../test/helpers → ../../helpers)
- vitest.config.ts include glob simplified back to 'test/**/*.test.ts'
- packages/proxy/src/cli/index.ts compliance check CLI command updated
- docs/compliance/*.md, CLAUDE.md, OPERATIONS.md path references updated

693 tests still pass; typecheck green.
