+++ docs/compliance/atlas-mapping.md
@@ -4,14 +4,14 @@ Aquaman maps to **MITRE ATLAS v5.3 (Adversarial Threat Landscape for Artificial-
 
 Below: each ATLAS technique aquaman addresses, the aquaman feature that mitigates it, and the runnable test that proves the mitigation holds.
 
-Run all ATLAS conformance tests with `aquaman compliance check` (emits a JSON evidence report) or `vitest run tests/compliance/atlas/`.
+Run all ATLAS conformance tests with `aquaman compliance check` (emits a JSON evidence report) or `vitest run test/compliance/atlas/`.
 
 | Technique | Title | Aquaman mitigation | Test |
 |---|---|---|---|
-| **AML.T0055** | Unsecured Credentials | Credentials live in a vault backend (Keychain, 1Password, HashiCorp Vault, KeePassXC, encrypted-file, systemd-creds, Bitwarden). The agent process never sees them — only a sentinel base-URL and a placeholder `aquaman-proxy-managed` marker. The proxy strips placeholder + injects the real credential before forwarding upstream. Post-tool-output redactor catches secrets that might leak into transcripts. | [`tests/compliance/atlas/t0055-unsecured-credentials.test.ts`](../../tests/compliance/atlas/t0055-unsecured-credentials.test.ts) |
-| **AML.T0012** | Valid Accounts | Even with valid credentials injected, request-level policy enforcement (method + path deny rules per service) blocks privileged paths *before* credentials are injected. Hash-chained audit log records every denied request with tamper-evident chain. Denied requests never reach the upstream provider. | [`tests/compliance/atlas/t0012-valid-accounts.test.ts`](../../tests/compliance/atlas/t0012-valid-accounts.test.ts) |
-| **AML.T0062** | Exfiltration via AI Agent Tool Invocation | Same policy engine + audit log. An agent that tries to exfiltrate via a sanctioned tool call (e.g., `Authorization`-bearing `curl`) is intercepted by the proxy, evaluated against policy, and either denied (403) or logged. Header-stripping in the HTTP interceptor removes `Authorization` + `X-API-Key` from outgoing requests before routing — so even if the agent presets them, they're scrubbed before egress. | [`tests/compliance/atlas/t0062-exfiltration-via-tool.test.ts`](../../tests/compliance/atlas/t0062-exfiltration-via-tool.test.ts) |
-| **AML.T0090** | OS Credential Dumping (Process credential access) | Credentials never reside in the agent's process address space. They live in a separate proxy process accessed via Unix Domain Socket (`chmod 0o600`). Even RCE in the agent cannot dump credentials from the agent's own memory because they were never there. The broker endpoint (v0.12.0+) materializes credentials per tool call and the consumer scrubs them after the indicated TTL. | [`tests/compliance/atlas/t0090-os-credential-dumping.test.ts`](../../tests/compliance/atlas/t0090-os-credential-dumping.test.ts) |
+| **AML.T0055** | Unsecured Credentials | Credentials live in a vault backend (Keychain, 1Password, HashiCorp Vault, KeePassXC, encrypted-file, systemd-creds, Bitwarden). The agent process never sees them — only a sentinel base-URL and a placeholder `aquaman-proxy-managed` marker. The proxy strips placeholder + injects the real credential before forwarding upstream. Post-tool-output redactor catches secrets that might leak into transcripts. | [`test/compliance/atlas/t0055-unsecured-credentials.test.ts`](../../test/compliance/atlas/t0055-unsecured-credentials.test.ts) |
+| **AML.T0012** | Valid Accounts | Even with valid credentials injected, request-level policy enforcement (method + path deny rules per service) blocks privileged paths *before* credentials are injected. Hash-chained audit log records every denied request with tamper-evident chain. Denied requests never reach the upstream provider. | [`test/compliance/atlas/t0012-valid-accounts.test.ts`](../../test/compliance/atlas/t0012-valid-accounts.test.ts) |
+| **AML.T0062** | Exfiltration via AI Agent Tool Invocation | Same policy engine + audit log. An agent that tries to exfiltrate via a sanctioned tool call (e.g., `Authorization`-bearing `curl`) is intercepted by the proxy, evaluated against policy, and either denied (403) or logged. Header-stripping in the HTTP interceptor removes `Authorization` + `X-API-Key` from outgoing requests before routing — so even if the agent presets them, they're scrubbed before egress. | [`test/compliance/atlas/t0062-exfiltration-via-tool.test.ts`](../../test/compliance/atlas/t0062-exfiltration-via-tool.test.ts) |
+| **AML.T0090** | OS Credential Dumping (Process credential access) | Credentials never reside in the agent's process address space. They live in a separate proxy process accessed via Unix Domain Socket (`chmod 0o600`). Even RCE in the agent cannot dump credentials from the agent's own memory because they were never there. The broker endpoint (v0.12.0+) materializes credentials per tool call and the consumer scrubs them after the indicated TTL. | [`test/compliance/atlas/t0090-os-credential-dumping.test.ts`](../../test/compliance/atlas/t0090-os-credential-dumping.test.ts) |
 
 ## What this is not
