+++ docs/compliance/nist-800-53.md
@@ -4,15 +4,15 @@ Aquaman maps to **NIST SP 800-53 Rev 5** control families that govern authentica
 
 | Control ID | Family | Aquaman feature | Test |
 |---|---|---|---|
-| **IA-5** | Authenticator Management | Seven vault backend implementations (Keychain, 1Password, HashiCorp Vault, KeePassXC, encrypted-file, systemd-creds, Bitwarden). Credentials are stored, rotated, and accessed via the vault's native APIs — aquaman never persists them itself. | [`tests/compliance/nist/ia-5-authenticator-mgmt.test.ts`](../../tests/compliance/nist/ia-5-authenticator-mgmt.test.ts) |
-| **AC-3** | Access Enforcement | Per-service request-policy engine evaluates method + path against allow/deny rules *before* credentials are injected. Denied requests return 403 and never reach the upstream provider. | unit: [`tests/compliance/nist/ac-3-access-enforcement.test.ts`](../../tests/compliance/nist/ac-3-access-enforcement.test.ts) + integration via [`tests/compliance/atlas/t0012-valid-accounts.test.ts`](../../tests/compliance/atlas/t0012-valid-accounts.test.ts) (full proxy round-trip asserting 403) |
+| **IA-5** | Authenticator Management | Seven vault backend implementations (Keychain, 1Password, HashiCorp Vault, KeePassXC, encrypted-file, systemd-creds, Bitwarden). Credentials are stored, rotated, and accessed via the vault's native APIs — aquaman never persists them itself. | [`test/compliance/nist/ia-5-authenticator-mgmt.test.ts`](../../test/compliance/nist/ia-5-authenticator-mgmt.test.ts) |
+| **AC-3** | Access Enforcement | Per-service request-policy engine evaluates method + path against allow/deny rules *before* credentials are injected. Denied requests return 403 and never reach the upstream provider. | unit: [`test/compliance/nist/ac-3-access-enforcement.test.ts`](../../test/compliance/nist/ac-3-access-enforcement.test.ts) + integration via [`test/compliance/atlas/t0012-valid-accounts.test.ts`](../../test/compliance/atlas/t0012-valid-accounts.test.ts) (full proxy round-trip asserting 403) |
 | **AC-6** | Least Privilege | The broker endpoint (v0.12.0+) materializes credentials per tool call with a short-lived TTL (default 60s, max 3600s). The agent's shell environment holds the credential only for the duration of one command, not for the agent process lifetime. | covered by `AC-3` test + broker tests in `test/e2e/credential-proxy.test.ts` |
 | **SC-12** | Cryptographic Key Establishment & Management | Vault backends are operator-chosen and operator-managed. Aquaman delegates key-management responsibilities to the chosen backend (Keychain, Vault, etc.) rather than rolling its own. | covered by `IA-5` test |
-| **SC-28** | Protection of Information at Rest | Audit log is written `chmod 0o600` to `~/.aquaman/audit/current.jsonl`. Credentials at rest live in the vault backend (operator-chosen encryption). Aquaman's own writes (auth-profiles.json, config.yaml) use `0o600`. | [`tests/compliance/nist/sc-28-protection-at-rest.test.ts`](../../tests/compliance/nist/sc-28-protection-at-rest.test.ts) |
-| **AU-2** | Audit Events | Every credential access and policy decision produces an audit event with `RequestInfo` carrying id, service, method, path, statusCode, authenticated. | [`tests/compliance/nist/au-10-tamper-evident.test.ts`](../../tests/compliance/nist/au-10-tamper-evident.test.ts) (same fixture covers AU-2/9/10) |
-| **AU-9** | Protection of Audit Information | Audit log is hash-chained (every entry contains the SHA-256 of the previous entry plus its own data). Tampering with any entry invalidates the chain from that point forward. Log is `chmod 0o600`. | [`tests/compliance/nist/au-10-tamper-evident.test.ts`](../../tests/compliance/nist/au-10-tamper-evident.test.ts) |
-| **AU-10** | Non-Repudiation | The hash chain provides non-repudiation: any mutation, insertion, or deletion of an entry breaks the chain and is detectable via `AuditLogger.verifyIntegrity()`. | [`tests/compliance/nist/au-10-tamper-evident.test.ts`](../../tests/compliance/nist/au-10-tamper-evident.test.ts) |
-| **SI-10** | Information Input Validation | Secret-pattern redactor scrubs known credential shapes (AWS, GitHub, Stripe, Slack, OpenAI, Anthropic, JWTs, PEM private keys, …) from agent inputs/outputs. Service names + key names validated against safe regex before vault lookup. | [`tests/compliance/nist/si-10-input-validation.test.ts`](../../tests/compliance/nist/si-10-input-validation.test.ts) |
+| **SC-28** | Protection of Information at Rest | Audit log is written `chmod 0o600` to `~/.aquaman/audit/current.jsonl`. Credentials at rest live in the vault backend (operator-chosen encryption). Aquaman's own writes (auth-profiles.json, config.yaml) use `0o600`. | [`test/compliance/nist/sc-28-protection-at-rest.test.ts`](../../test/compliance/nist/sc-28-protection-at-rest.test.ts) |
+| **AU-2** | Audit Events | Every credential access and policy decision produces an audit event with `RequestInfo` carrying id, service, method, path, statusCode, authenticated. | [`test/compliance/nist/au-10-tamper-evident.test.ts`](../../test/compliance/nist/au-10-tamper-evident.test.ts) (same fixture covers AU-2/9/10) |
+| **AU-9** | Protection of Audit Information | Audit log is hash-chained (every entry contains the SHA-256 of the previous entry plus its own data). Tampering with any entry invalidates the chain from that point forward. Log is `chmod 0o600`. | [`test/compliance/nist/au-10-tamper-evident.test.ts`](../../test/compliance/nist/au-10-tamper-evident.test.ts) |
+| **AU-10** | Non-Repudiation | The hash chain provides non-repudiation: any mutation, insertion, or deletion of an entry breaks the chain and is detectable via `AuditLogger.verifyIntegrity()`. | [`test/compliance/nist/au-10-tamper-evident.test.ts`](../../test/compliance/nist/au-10-tamper-evident.test.ts) |
+| **SI-10** | Information Input Validation | Secret-pattern redactor scrubs known credential shapes (AWS, GitHub, Stripe, Slack, OpenAI, Anthropic, JWTs, PEM private keys, …) from agent inputs/outputs. Service names + key names validated against safe regex before vault lookup. | [`test/compliance/nist/si-10-input-validation.test.ts`](../../test/compliance/nist/si-10-input-validation.test.ts) |
 
 ## SOC 2 Common Criteria mapping
